Moved Manager Classes to their own package. TODO, rename ConfigManager.java to Config.java
