+++ gradle.properties
@@ -8,7 +8,7 @@ org.gradle.jvmargs=-Xmx1G
 	loader_version=0.11.3
 
 # Mod Properties
-	mod_version = 1.0.0.13-ss21w15f
+	mod_version = 1.0.0.14-ss21w15g
 	maven_group = zetamod
 	archives_base_name = zetamod-beta
 	# portalapi = 0.0.1-beta22-1.17
