+++ src/main/java/zetamod/mod/managers/GeneralManager.java
@@ -1,8 +1,7 @@
-package zetamod.mod.features;
+package zetamod.mod.managers;
 
 import me.zeroeightsix.fiber.exception.FiberException;
 import net.fabricmc.api.ModInitializer;
-import zetamod.mod.main.ConfigManager;
 import zetamod.mod.main.ZetaModMain;
 import net.minecraft.util.math.MathHelper;
 import org.apache.logging.log4j.Level;
