+++ src/main/java/zetamod/mod/mixins/FarLandsMixins.java
@@ -1,6 +1,6 @@
 package zetamod.mod.mixins;
 
-import zetamod.mod.features.GeneralManager;
+import zetamod.mod.managers.GeneralManager;
 import net.minecraft.util.math.noise.OctavePerlinNoiseSampler;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Overwrite;
