+++ src/main/java/zetamod/mod/mixins/NoiseMixinTemporary.java
@@ -3,7 +3,7 @@ package zetamod.mod.mixins;
 import net.minecraft.world.gen.chunk.NoiseSamplingConfig;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
-import zetamod.mod.main.ConfigManager;
+import zetamod.mod.managers.ConfigManager;
 
 @Mixin(NoiseSamplingConfig.class)
 public class NoiseMixinTemporary {
